# gcp/projects/cuttlefish-hackathon/hackathon-controller/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hackathon Controller",
    "description": "API that validates Google Cloud Tokens and stores requests in Firestore",
    "version": "0.1.0"
  },
  "paths": {
    "/api/principals": {
      "get": {
        "summary": "List Principals",
        "description": "Lists out the registered principals from the database.",
        "operationId": "list_principals_api_principals_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Principal",
        "description": "Stores the metadata in a lightweight Firebase (Firestore) database.\nIAP handles authentication securely.",
        "operationId": "add_principal_api_principals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrincipalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/principals/{doc_id}/grant_permissions": {
      "post": {
        "summary": "Grant Permissions",
        "description": "Grants the Compute Engine default service account object viewer permissions to the bucket.\nIAP handles authentication securely.",
        "operationId": "grant_permissions_api_principals__doc_id__grant_permissions_post",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/principals/{doc_id}": {
      "patch": {
        "summary": "Update Principal",
        "description": "Updates an existing principal to set their nickname.\nIAP handles authentication securely.",
        "operationId": "update_principal_api_principals__doc_id__patch",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrincipalUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Principal",
        "description": "Deletes an existing principal.\nIAP handles authentication securely.",
        "operationId": "delete_principal_api_principals__doc_id__delete",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PrincipalRequest": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "project"
        ],
        "title": "PrincipalRequest"
      },
      "PrincipalUpdateRequest": {
        "properties": {
          "nickname": {
            "type": "string",
            "title": "Nickname"
          }
        },
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "PrincipalUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}